+++ lib/adjust_event.dart
@@ -10,13 +10,15 @@ import 'dart:convert';
 
 class AdjustEvent {
   String _eventToken;
-  num? _revenue;
   String? _currency;
+  String? receipt;
+  String? transactionId;
+  String? productId;
+  String? purchaseToken;
+  String? callbackId;
+  num? _revenue;
   Map<String, String>? _callbackParameters;
   Map<String, String>? _partnerParameters;
-
-  String? transactionId;
-  String? callbackId;
 
   AdjustEvent(this._eventToken) {
     _callbackParameters = new Map<String, String>();
@@ -47,6 +49,15 @@ class AdjustEvent {
     }
     if (transactionId != null) {
       eventMap['transactionId'] = transactionId;
+    }
+    if (receipt != null) {
+      eventMap['receipt'] = receipt;
+    }
+    if (productId != null) {
+      eventMap['productId'] = productId;
+    }
+    if (purchaseToken != null) {
+      eventMap['purchaseToken'] = purchaseToken;
     }
     if (callbackId != null) {
       eventMap['callbackId'] = callbackId;