+++ test/app/lib/main.dart
@@ -20,6 +20,7 @@ class _MyAppState extends State<MyApp> {
   String? _baseUrl;
   String? _gdprUrl;
   String? _subscriptionUrl;
+  String? _purchaseVerificationUrl;
   late String _controlUrl;
   late CommandExecutor _commandExecutor;
 
@@ -27,13 +28,14 @@ class _MyAppState extends State<MyApp> {
   void initState() {
     super.initState();
 
-    String _address = '192.168.8.184';
+    String _address = '192.168.2.112';
     if (Platform.isAndroid) {
       String _protocol = 'https';
       String _port = '8443';
       _baseUrl = _protocol + '://' + _address + ':' + _port;
       _gdprUrl = _protocol + '://' + _address + ':' + _port;
       _subscriptionUrl = _protocol + '://' + _address + ':' + _port;
+      _purchaseVerificationUrl = _protocol + '://' + _address + ':' + _port;
       _controlUrl = 'ws://' + _address + ':1987';
     } else {
       String _protocol = 'http';
@@ -41,12 +43,13 @@ class _MyAppState extends State<MyApp> {
       _baseUrl = _protocol + '://' + _address + ':' + _port;
       _gdprUrl = _protocol + '://' + _address + ':' + _port;
       _subscriptionUrl = _protocol + '://' + _address + ':' + _port;
+      _purchaseVerificationUrl = _protocol + '://' + _address + ':' + _port;
       _controlUrl = 'ws://' + _address + ':1987';
     }
 
     // Initialise command executor.
     _commandExecutor =
-        new CommandExecutor(_baseUrl, _gdprUrl, _subscriptionUrl);
+        new CommandExecutor(_baseUrl, _gdprUrl, _subscriptionUrl, _purchaseVerificationUrl);
 
     // Initialise test library.
     TestLib.setExecuteCommandHalder((final dynamic callArgs) {